+++ src/js/app/index.js
@@ -7,15 +7,10 @@ import { HistoryMiddleware, Provider } from 'react-redux';
 import { Router, useRouterHistory } from 'react-router';
 import createBrowserHistory from 'history/lib/createBrowserHistory';
 import { syncHistoryWithStore } from 'react-router-redux';
-import injectTapEventPlugin from 'react-tap-event-plugin';
 import configureStore from './store';
 import getRoutes from './routes';
 import { sanitizeContextRoot } from './utils/url-helper';
 import './devtools/chrome-react-perf';
-// import '../../scss/index.scss';
-
-// http://www.material-ui.com/#/get-started/installation
-injectTapEventPlugin();
 
 // instead of using `browserHistory` from react-router, create one
 // with basename to allow app to specify different context root